Use trampolines for RELOC_REL32 > 2GiB
relocate function now takes a pointer to the jmptbl page allocated in the
image being loaded. If a RELOC_REL32 relocation is encountered where the
offset is greater than 2GiB then 16 bytes of the page is used to generate
a trampoline to the function.

The new functionality is exposed as flexdll_relocate_v2 and passed in the
FLEXDLL_RELOCATE_V2 environment variable. The old flexdll_relocate pointer
is still passed in FLEXDLL_RELOCATE by flexdll_dlopen meaning that an
executable compiled with this new version can still load DLLs compiled
with older versions (these will simply fail if the relocations are too far
away, as before). Similarly, flexdll_init falls back to FLEXDLL_RELOCATE
if FLEXDLL_RELOCATE_V2 is not set meaning that a DLL compiled with this
new version can also be loaded by an executable compiled with older
versions.

The symbol table now has to be copied to be augmented with an extra
pointer for the trampoline, as the symtbl itself cannot be changed without
breaking this backwards compatibility.